organization of resources into more specific folders
